Enhanced C64 port of John Jainschigg's "Portrait of Liberty"
Original source was published July 1986 in Family Computing Magazine.
This port reimplements it in machine language, using faster bitmap
libraries and the three-voice music from the PCjr source. It also
implements a floodfill operation so that the graphical display is
closer to the PC's display than the original C64 display.