Initial, most direct implementation of SimEvo.
This is a relatively close adaptation of A.K. Dewdney's description
of the program "Simulated Evolution" from his Computer Recreations
column, first published in Scientific American in May 1989.

This first implementation runs only on the UNIX commandline and simply
reports the life events and genomes of the bugs as they evolve. A
number of optimizations and simplifications have also been introduced
as a demonstration that they would not warp the simulation unduly while
leaving the implementation more friendly to 16-bit math.